Fiber增加了一个参数，可以传一个rootId进去
这样作用是 1. 可以让MapCopy.Id等于MapId 2省去了把MapCopy.Id拼在MapName中

+++ Packages/cn.etetet.mapmanager/Scripts/Model/Server/MergeLinesComponent.cs
@@ -4,8 +4,8 @@ namespace ET.Server
 {
     public struct MergeLineInfo
     {
-        public int LineNum1;
-        public int LineNum2;
+        public long LineNum1;
+        public long LineNum2;
         public long Time;
     }
